+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Latest
+* Added removeCommsCallback to HeLMS
+* Added optional static object respawns/despawns to the zonal spawning methods
+* Fixed undefined constant in asset_pools
 ## 1.9a
 * Improved KOTH scoreboard and added disappearing smoke workaround
 * Added steersman zig-zag enable/disable function