+++ Changelog.md
@@ -4,6 +4,7 @@
 * Added removeCommsCallback to HeLMS
 * Added optional static object respawns/despawns to the zonal spawning methods
 * Fixed undefined constant in asset_pools
+* Added support for non-circular zones for HeLMS zone-based respawns
 ## 1.9a
 * Improved KOTH scoreboard and added disappearing smoke workaround
 * Added steersman zig-zag enable/disable function