+++ Changelog.md
@@ -8,6 +8,7 @@
 * Added `helms.ai` and `helms.ai.setAlarmState`
 * KOTH 1.3: Kill detection rule improvements
 * KOTH 1.3: Crown spawn zone can be a multiple of the zone of play
+* KOTH 1.3: Fixed inability to create multiple games with the same name
 ## 1.9a
 * Improved KOTH scoreboard and added disappearing smoke workaround
 * Added steersman zig-zag enable/disable function